+++ Editor/Scripts/IPSceneGUI.cs
@@ -15,6 +15,9 @@ namespace InstancePainter.Editor
 
         public static void DrawGUI(SceneView p_sceneView)
         {
+            if (EditorApplication.isCompiling || BuildPipeline.isBuildingPlayer)
+                return;
+            
             Handles.BeginGUI();
 
             var rect = p_sceneView.camera.GetScaledPixelRect();